Added condition-command scheme in telegram bot

+++ src/main/java/com/github/graywizard123/tgbot/db/models/Meal.java
@@ -5,19 +5,19 @@ import com.github.graywizard123.tgbot.db.table.TableColumnType;
 
 public class Meal {
 
-    private final int id;
+    private final long id;
     private String name;
     private String description;
     private int price;
 
-    public Meal(int id, String name, String description, int price) {
+    public Meal(long id, String name, String description, int price) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.price = price;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
